completed auth

+++ server/controllers/shift.controller.js
@@ -3,11 +3,7 @@ const db = require('../models');
 
 exports.getAllShifts = async (req, res) => {
   try {
-    const { userId } = req.params;
-    console.log('get all Shifts ******************', userId);
-    let shifts = await db.Shift.findAll({
-      where: { user_id: userId }
-    });
+    let shifts = await db.Shift.findAll();
     res
       .status(200)
       .send(shifts);
@@ -20,13 +16,10 @@ exports.getAllShifts = async (req, res) => {
 
 exports.addShift = async (req, res) => {
   try {
-    const { userId } = req.params;
-    console.log('Add Shift  ******************', userId);
     let newShift = await db.Shift.create({
       day_number: req.body.day_number,
       people_required: req.body.people_required,
       shift_type_id: req.body.shift_type_id,
-      user_id: userId
     });
     res
       .status(201)
@@ -40,12 +33,9 @@ exports.addShift = async (req, res) => {
 
 exports.deleteShift = async (req, res) => {
   let id = req.params.id;
-  const { userId } = req.params;
-  console.log('DeleteShift ******************', userId);
   try {
-    await db.Shift.destroy({
-      where: { shift_id: id, user_id: userId }
-    });
+    await db.Shift.destroy();
+    await db.Shift.destroy({ where: { shift_id: id, user_id: userId } });
     res
       .status(200)
       .send({ message: `Shift deleted successfully` });
@@ -59,11 +49,12 @@ exports.deleteShift = async (req, res) => {
 
 exports.updateShift = async (req, res) => {
   let id = req.params.id;
-  const { userId } = req.params;
-  console.log('Update Shift ******************', userId);
   try {
     let toBeUpdatedArr = await db.Shift.findAll({
-      where: { shift_id: id, user_id: userId }
+      where: {
+        shift_id: id,
+        // user_id: userId
+      }
     });
     let temp = toBeUpdatedArr[0];
     await temp.set({ ...req.body });
